Release script update

+++ release.sh
@@ -1,10 +1,18 @@
 # Script to release the mod into a zip file.
 # Assumes BABUN on windows or something like that on Linux (duh)
 
+# Stop on error.
+set -e
+
+if [ $# -eq 0 ] ; then
+    echo "Provide release number"
+    exit 1
+fi
+
 REL=$1
 OFNAME=yupgi_alert_r${REL}.zip
-
 PREFIX=`pwd`/tmp/yupgi_alert
+
 rm -rf tmp
 rm -f $OFNAME.zip
 mkdir -p $PREFIX
@@ -17,8 +25,6 @@ rm -rf $PREFIX/oramod/{.git,assets,.gitattributes}
 rm -f $PREFIX/oramod/{.gitignore,release.sh}
 rm -f $PREFIX/oramod/rules/_buildpal_order.py
 rm -f $PREFIX/oramod/rules/_infest.py
-# Remove maps to prevent crash haha
-#rm -rf $PREFIX/oramod/maps
 
 # copy the DLL file and the license.
 echo "Copying DLL files and license info"
@@ -29,7 +35,7 @@ cp ../../OpenRA.Game.exe $PREFIX/OpenRA.yupgi_alert.exe
 cp ../../OpenRA.Mods.yupgi_alert/{LICENSE.AS,AUTHORS.AS} $PREFIX
 
 # patch mod.yaml
-mv $PREFIX/oramod/mod.yaml.release $PREFIX/oramod/mod.yaml
+python3 make_mod_yaml.py mod.yaml $REL > $PREFIX/oramod/mod.yaml
 
 # now in tmp dir,
 echo "Archiving into $OFNAME"
